requirements.txt
Django==1.10
wsgiref==0.1.2
geopy
django-allauth
unipath==1.1
djangorestframework
markdown
django-filter
psycopg2
django-rest-swagger==0.3.8
django-pagseguro2
